double -> int for zone in OGRSpatialReference::importFromESRI().
Accidentally converted zone to a double when cleanup up ogr_srs_esri.cpp in r33444.


git-svn-id: https://svn.osgeo.org/gdal/trunk@33445 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srs_esri.cpp
@@ -712,7 +712,7 @@ OGRErr OGRSpatialReference::importFromESRI( char **papszPrj )
             const double dfCentralMeridian = OSR_GDV( papszPrj, "PARAM_1", 0.0 );
             const double dfRefLat = OSR_GDV( papszPrj, "PARAM_2", 0.0 );
 
-            const double nZone = static_cast<int>(
+            const int nZone = static_cast<int>(
                 (dfCentralMeridian+183) / 6.0 + 0.0000001 );
             SetUTM( nZone, dfRefLat >= 0.0 );
         }
